Clean up Dart port.
