+++ sql/mysql/Oracle/Dart/SqlMode.dart
@@ -1,0 +1,9 @@
+/** SQL modes that control parsing behavior. */
+enum SqlMode {
+  noMode,
+  ansiQuotes,
+  highNotPrecedence,
+  pipesAsConcat,
+  ignoreSpace,
+  noBackslashEscapes,
+}
